Move to VCPKG.
