+++ src/bsafile.cpp
