+++ src/bsafolder.cpp
