--- src/bsatk.h
@@ -1,15 +1,0 @@
-/*
- * File:   bsatk.h
- * Author: tannin
- *
- * Created on March 16, 2012, 10:35 AM
- */
-
-#ifndef BSATK_H
-#define BSATK_H
-
-#include "bsaarchive.h"
-#include "bsafile.h"
-#include "bsafolder.h"
-
-#endif /* BSATK_H */
