+++ src/filehash.cpp
