+++ vcpkg-configuration.json
@@ -1,0 +1,21 @@
+{
+  "default-registry": {
+    "kind": "git",
+    "repository": "https://github.com/Microsoft/vcpkg",
+    "baseline": "f61a294e765b257926ae9e9d85f96468a0af74e7"
+  },
+  "registries": [
+    {
+      "kind": "git",
+      "repository": "https://github.com/Microsoft/vcpkg",
+      "baseline": "f61a294e765b257926ae9e9d85f96468a0af74e7",
+      "packages": ["boost*", "boost-*"]
+    },
+    {
+      "kind": "git",
+      "repository": "https://github.com/ModOrganizer2/vcpkg-registry",
+      "baseline": "a07812dd075d7ce5586e5cc9052050f229e6ecb0",
+      "packages": ["mo2-dds-header"]
+    }
+  ]
+}
